Fix hurl not minimizing to tray after rule match
This only happens when hurl is not in tray when it's launched and the
launched url matches with a rule. This keeps hurl in the taskbar and
sometimes takes the focus from the browser that rule match launched.

fixed by launching the browser only after making sure hurl is
successfully launched.

Also refactor AutoRuleCheck to support doing this

closes #193 and #179